requirements.txt
Flask
SpeechRecognition
google-generativeai
pandas
icalendar
python-dotenv